Open restore settings for coinManager. * Add styled for View (move informed + copyable). * Refactor code.
