+++ UnstoppableWallet/UnstoppableWallet/Core/Adapters/StellarSendHelper.swift
@@ -154,7 +154,7 @@ extension StellarSendHelper {
 
             viewItems.append(
                 .fee(
-                    title: SendField.InformedTitle("fee_settings.network_fee".localized, info: .fee),
+                    title: ComponentInformedTitle("fee_settings.network_fee".localized, info: .fee),
                     amountData: .init(appValue: appValue, currencyValue: currencyValue)
                 )
             )
