+++ UnstoppableWallet/UnstoppableWallet/Core/Adapters/TonSendHelper.swift
@@ -135,7 +135,7 @@ extension TonSendHelper {
 
             viewItems.append(
                 .fee(
-                    title: SendField.InformedTitle("fee_settings.network_fee".localized, info: .fee),
+                    title: ComponentInformedTitle("fee_settings.network_fee".localized, info: .fee),
                     amountData: .init(appValue: appValue, currencyValue: currencyValue)
                 )
             )
