+++ UnstoppableWallet/UnstoppableWallet/Core/Adapters/ZanoSendHelper.swift
@@ -40,7 +40,7 @@ class ZanoSendHelper {
 
         return [
             .fee(
-                title: SendField.InformedTitle("fee_settings.network_fee".localized, info: .fee),
+                title: ComponentInformedTitle("fee_settings.network_fee".localized, info: .fee),
                 amountData: .init(appValue: appValue, currencyValue: currencyValue)
             ),
         ]
