+++ UnstoppableWallet/UnstoppableWallet/Modules/MultiSwap/SwapFinalQuote.swift
@@ -68,7 +68,7 @@ class SwapFinalQuote {
 
         if let estimatedTime {
             fields.append(.simpleValue(
-                title: SendField.InformedTitle("swap.swapped_time".localized, info: InfoDescription(
+                title: ComponentInformedTitle("swap.swapped_time".localized, info: InfoDescription(
                     title: "swap.swapped_time".localized,
                     description: "swap.swapped_time.info".localized
                 )),
