+++ UnstoppableWallet/UnstoppableWallet/Modules/SendNew/TonSendHandler.swift
@@ -201,7 +201,7 @@ extension TonSendHandler {
 
                 viewItems.append(
                     .fee(
-                        title: SendField.InformedTitle("fee_settings.network_fee".localized, info: .fee),
+                        title: ComponentInformedTitle("fee_settings.network_fee".localized, info: .fee),
                         amountData: .init(appValue: appValue, currencyValue: currencyValue)
                     )
                 )
