+++ UnstoppableWallet/UnstoppableWallet/Modules/SendNew/UtxoSendHelper.swift
@@ -19,7 +19,7 @@ class UtxoSendHelper {
 
         return [
             .fee(
-                title: SendField.InformedTitle("fee_settings.network_fee".localized, info: .fee),
+                title: ComponentInformedTitle("fee_settings.network_fee".localized, info: .fee),
                 amountData: amountData
             ),
         ]
