+++ UnstoppableWallet/UnstoppableWallet/Modules/SendNew/ZcashSendHandler.swift
@@ -126,7 +126,7 @@ extension ZcashSendHandler {
 
                 feeFields.append(
                     .fee(
-                        title: SendField.InformedTitle("fee_settings.network_fee".localized, info: .fee),
+                        title: ComponentInformedTitle("fee_settings.network_fee".localized, info: .fee),
                         amountData: .init(appValue: appValue, currencyValue: currencyValue)
                     )
                 )
