+++ UnstoppableWallet/UnstoppableWallet/Modules/ZCashShieldTokens/ShieldSendHandler.swift
@@ -129,7 +129,7 @@ extension ShieldSendHandler {
                 .init(flowFields, isFlow: true),
                 .init(fields + [
                     .fee(
-                        title: SendField.InformedTitle("fee_settings.network_fee".localized, info: .fee),
+                        title: ComponentInformedTitle("fee_settings.network_fee".localized, info: .fee),
                         amountData: .init(appValue: appValue, currencyValue: currencyValue)
                     ),
                 ], isMain: false),
